+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-03-12
+
 ### Added
 
 - Support for single and double quoted map keys in TableTest maps.
@@ -149,7 +151,8 @@
 - Syntax highlighting in TableTest tables
 - Auto-injection of TableTest language for Java `@TableTest` annotations
 
-[Unreleased]: https://github.com/nchaugen/tabletest-intellij/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/nchaugen/tabletest-intellij/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nchaugen/tabletest-intellij/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/nchaugen/tabletest-intellij/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nchaugen/tabletest-intellij/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/nchaugen/tabletest-intellij/compare/v0.2.1...v0.2.2